fjerner simulering fra urskog #536
