+++ apps/urskog/README.md
@@ -3,7 +3,6 @@
 I urskogen finner man kulturarv fra oldtiden.
 
 ## Features
-- Propagerer simuleringer fra Kafka til REST-clienten vår (helved-simulering)
 - Propagerer avstemminger fra Kafka til MQ
 - Propagerer oppdrag fra Kafka til MQ
 - Propagerer kvitteringer fra MQ til Kafka 
@@ -19,12 +18,6 @@ I urskogen finner man kulturarv fra oldtiden.
 - Status utledes og legges på "helved.status.v1" (200, 400 eller 500 type status)
 
 ![oppdrag](oppdrag.svg)
-
-- Leser helved.simulering.v1 og sender request via proxy til gandalf og os/ur
-- Svarene mappes til ulike formater og legges på fagsystemene sine respektive topics, pga sensitiv informasjon
-- Status returneres ved bad-request, eller ved feil mot OS/UR, gandalf, proxy etc.
-
-![simulering](simulering.svg)
 
 - Avstemminger opprettes i vedskiva, ved å se på alle oppdrag/kvitteringer.
 - Disse blir splittet opp i en start/data/slutt melding per fagsystem.
