+++ apps/urskog/prod.yml
@@ -81,16 +81,12 @@ spec:
           - name: "pgaudit.log_parameter"
             value: "on"
   envFrom:
-     # TODO: sjekk hva disse inneholder, er det samme eller noe ulikt?
-    - secret: utsjekk-oppdrag-simulering
     - secret: utsjekk-oppdrag
   env:
     - name: PROXY_HOST
       value: https://helved-ws-proxy.prod-fss-pub.nais.io
     - name: PROXY_SCOPE
       value: api://prod-fss.helved.ws-proxy/.default
-    - name: SIMULERING_PATH
-      value: cics/cics/services/simulerFpServiceWSBinding
     - name: MQ_HOSTNAME
       value: mpls02.adeo.no
     - name: MQ_PORT
