--- apps/urskog/test/simuler-fault-no-soap-body.xml
@@ -1,9 +1,0 @@
-<SOAP-ENV:Envelope xmlns:SOAP-ENV="http://schemas.xmlsoap.org/soap/envelope/">
-    <SOAP-ENV:Body>
-        <SOAP-ENV:Fault xmlns="">
-            <faultcode>SOAP-ENV:Server</faultcode>
-            <faultstring>The target application did not create a new SOAP Body</faultstring>
-            <detail/>
-        </SOAP-ENV:Fault>
-    </SOAP-ENV:Body>
-</SOAP-ENV:Envelope>
